+++ src/kagglehub/http_resolver.py
@@ -2,7 +2,6 @@ import logging
 import os
 import tarfile
 import zipfile
-from pathlib import Path
 from typing import Optional
 
 import requests
@@ -21,6 +20,7 @@ from kagglehub.handle import CompetitionHandle, DatasetHandle, ModelHandle, Note
 from kagglehub.resolver import Resolver
 
 DATASET_CURRENT_VERSION_FIELD = "currentVersionNumber"
+NOTEBOOK_CURRENT_VERSION_FIELD = "currentVersionNumber"
 
 MODEL_INSTANCE_VERSION_FIELD = "versionNumber"
 MAX_NUM_FILES_DIRECT_DOWNLOAD = 25
@@ -55,7 +55,9 @@ class CompetitionHttpResolver(Resolver[CompetitionHandle]):
             os.makedirs(os.path.dirname(out_path), exist_ok=True)
 
             try:
-                download_needed = api_client.download_file(url_path, out_path, h, cached_path)
+                download_needed = api_client.download_file(
+                    url_path, out_path, h, cached_path, extract_auto_compressed_file=True
+                )
             except requests.exceptions.ConnectionError:
                 if cached_path:
                     return cached_path
@@ -83,7 +85,6 @@ class CompetitionHttpResolver(Resolver[CompetitionHandle]):
                     os.remove(archive_path)
                 return cached_path
 
-            os.makedirs(out_path, exist_ok=True)
             _extract_archive(archive_path, out_path)
             os.remove(archive_path)
 
@@ -108,14 +109,14 @@ class DatasetHttpResolver(Resolver[DatasetHandle]):
         elif dataset_path and force_download:
             delete_from_cache(h, path)
 
-        url_path = _build_dataset_download_url_path(h)
+        url_path = _build_dataset_download_url_path(h, path)
         out_path = get_cached_path(h, path)
 
         # Create the intermediary directories
         if path:
             # Downloading a single file.
             os.makedirs(os.path.dirname(out_path), exist_ok=True)
-            api_client.download_file(url_path + "&file_name=" + path, out_path, h)
+            api_client.download_file(url_path, out_path, h, extract_auto_compressed_file=True)
         else:
             # TODO(b/345800027) Implement parallel download when < 25 files in databundle.
             # Downloading the full archived bundle.
@@ -124,9 +125,6 @@ class DatasetHttpResolver(Resolver[DatasetHandle]):
 
             # First, we download the archive.
             api_client.download_file(url_path, archive_path, h)
-
-            # Create the directory to extract the archive to.
-            os.makedirs(out_path, exist_ok=True)
 
             _extract_archive(archive_path, out_path)
 
@@ -154,14 +152,14 @@ class ModelHttpResolver(Resolver[ModelHandle]):
         elif model_path and force_download:
             delete_from_cache(h, path)
 
-        url_path = _build_download_url_path(h)
+        url_path = _build_model_download_url_path(h, path)
         out_path = get_cached_path(h, path)
 
         # Create the intermediary directories
         if path:
             # Downloading a single file.
             os.makedirs(os.path.dirname(out_path), exist_ok=True)
-            api_client.download_file(url_path + "/" + path, out_path, h)
+            api_client.download_file(url_path, out_path, h, extract_auto_compressed_file=True)
         else:
             # List the files and decide how to download them:
             # - <= 25 files: Download files in parallel
@@ -174,9 +172,6 @@ class ModelHttpResolver(Resolver[ModelHandle]):
 
                 # First, we download the archive.
                 api_client.download_file(url_path, archive_path, h)
-
-                # Create the directory to extract the archive to.
-                os.makedirs(out_path, exist_ok=True)
 
                 _extract_archive(archive_path, out_path)
 
@@ -208,46 +203,37 @@ class NotebookOutputHttpResolver(Resolver[NotebookHandle]):
     def __call__(self, h: NotebookHandle, path: Optional[str] = None, *, force_download: Optional[bool] = False) -> str:
         api_client = KaggleApiV1Client()
 
-        cached_response = load_from_cache(h, path)
-        if cached_response and not force_download:
-            return cached_response  # Already cached
-        elif cached_response and force_download:
+        if not h.is_versioned():
+            h.version = _get_current_version(api_client, h)
+
+        nb_path = load_from_cache(h, path)
+        if nb_path and not force_download:
+            return nb_path  # Already cached
+        elif nb_path and force_download:
             delete_from_cache(h, path)
 
-        download_url_root = f"kernels/output/download/{h.owner}/{h.notebook}"
-        output_root = Path(get_cached_path(h, path))
+        url_path = _build_notebook_download_url_path(h, path)
+        out_path = get_cached_path(h, path)
 
-        # List the files and decide how to download them:
-        # - <= 25 files: Download files in parallel
-        # > 25 files: Download the archive and uncompress
-        (files, has_more) = self._list_files(api_client, h) if not path else ([path], False)
-        if has_more:
-            # TODO(b/379761520): add support for .tar.gz archived downloads
-            logger.warning(
-                f"Too many files in {h} (capped at {MAX_NUM_FILES_DIRECT_DOWNLOAD}). "
-                "Unable to download notebook output."
-            )
-            return ""
+        if path:
+            os.makedirs(os.path.dirname(out_path), exist_ok=True)
+            api_client.download_file(url_path, out_path, h, extract_auto_compressed_file=True)
+        else:
+            # TODO(b/345800027) Implement parallel download when < 25 files in databundle.
+            # Downloading the full archived bundle.
+            archive_path = get_cached_archive_path(h)
+            os.makedirs(os.path.dirname(archive_path), exist_ok=True)
 
-        # Download files individually in parallel
-        def _inner_download_file(filepath: str) -> None:
-            download_url_path = f"{download_url_root}/{filepath}"
-            full_output_filepath = output_root / filepath
+            # First, we download the archive.
+            api_client.download_file(url_path, archive_path, h)
 
-            os.makedirs(os.path.dirname(full_output_filepath), exist_ok=True)
-            api_client.download_file(download_url_path, str(full_output_filepath), h)
+            _extract_archive(archive_path, out_path)
 
-        thread_map(
-            _inner_download_file,
-            files,
-            desc=f"Downloading {len(files)} files",
-            max_workers=8,  # Never use more than 8 threads in parallel to download files.
-        )
+            # Delete the archive
+            os.remove(archive_path)
 
         mark_as_complete(h, path)
-
-        # TODO(b/377510971): when notebook is a Kaggle utility script, update sys.path
-        return str(output_root)
+        return out_path
 
     def _list_files(self, api_client: KaggleApiV1Client, h: NotebookHandle) -> tuple[list[str], bool]:
         query = f"kernels/output/list/{h.owner}/{h.notebook}?page_size={MAX_NUM_FILES_DIRECT_DOWNLOAD}"
@@ -263,6 +249,9 @@ class NotebookOutputHttpResolver(Resolver[NotebookHandle]):
 
 
 def _extract_archive(archive_path: str, out_path: str) -> None:
+    # Create the directory to extract the archive to.
+    os.makedirs(out_path, exist_ok=True)
+
     logger.info("Extracting files...")
     if tarfile.is_tarfile(archive_path):
         with tarfile.open(archive_path) as f:
@@ -292,6 +281,14 @@ def _get_current_version(api_client: KaggleApiV1Client, h: ResourceHandle) -> in
 
         return json_response[DATASET_CURRENT_VERSION_FIELD]
 
+    elif isinstance(h, NotebookHandle):
+        json_response = api_client.get(_build_get_notebook_url_path(h), h)
+        if NOTEBOOK_CURRENT_VERSION_FIELD not in json_response["metadata"]:
+            msg = f"Invalid GetKernel API response. Expected to include a {NOTEBOOK_CURRENT_VERSION_FIELD} field"
+            raise ValueError(msg)
+
+        return json_response["metadata"][NOTEBOOK_CURRENT_VERSION_FIELD]
+
     else:
         msg = f"Invalid ResourceHandle type {h}"
         raise ValueError(msg)
@@ -316,8 +313,12 @@ def _build_get_instance_url_path(h: ModelHandle) -> str:
     return f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/get"
 
 
-def _build_download_url_path(h: ModelHandle) -> str:
-    return f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/{h.version}/download"
+def _build_model_download_url_path(h: ModelHandle, path: Optional[str]) -> str:
+    base_url = f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/{h.version}/download"
+    if path:
+        return f"{base_url}/{path}"
+    else:
+        return base_url
 
 
 def _build_list_model_instance_version_files_url_path(h: ModelHandle) -> str:
@@ -329,8 +330,32 @@ def _build_get_dataset_url_path(h: DatasetHandle) -> str:
     return f"datasets/view/{h.owner}/{h.dataset}"
 
 
-def _build_dataset_download_url_path(h: DatasetHandle) -> str:
-    return f"datasets/download/{h.owner}/{h.dataset}?dataset_version_number={h.version}"
+def _build_get_notebook_url_path(h: NotebookHandle) -> str:
+    return f"kernels/pull?user_name={h.owner}&kernel_slug={h.notebook}"
+
+
+def _build_dataset_download_url_path(h: DatasetHandle, path: Optional[str]) -> str:
+    if not h.is_versioned():
+        msg = "No version provided"
+        raise ValueError(msg)
+
+    base_url = f"datasets/download/{h.owner}/{h.dataset}?dataset_version_number={h.version}"
+    if path:
+        return f"{base_url}&file_name={path}"
+    else:
+        return base_url
+
+
+def _build_notebook_download_url_path(h: NotebookHandle, path: Optional[str]) -> str:
+    if not h.is_versioned():
+        msg = "No version provided"
+        raise ValueError(msg)
+
+    base_url = f"kernels/output/download/{h.owner}/{h.notebook}?version_number={h.version}"
+    if path:
+        return f"{base_url}&file_path={path}"
+    else:
+        return base_url
 
 
 def _build_competition_download_all_url_path(h: CompetitionHandle) -> str: