+++ tests/test_http_dataset_download.py
@@ -1,5 +1,4 @@
 import os
-import zipfile
 from typing import Optional
 
 import kagglehub
@@ -9,13 +8,18 @@ from tests.fixtures import BaseTestCase
 
 from .server_stubs import dataset_download_stub as stub
 from .server_stubs import serv
-from .utils import create_test_cache
+from .utils import AUTO_COMPRESSED_FILE_NAME, create_test_cache
 
 INVALID_ARCHIVE_DATASET_HANDLE = "invalid/invalid/invalid/invalid/invalid"
 VERSIONED_DATASET_HANDLE = "sarahjeffreson/featured-spotify-artiststracks-with-metadata/versions/2"
 UNVERSIONED_DATASET_HANDLE = "sarahjeffreson/featured-spotify-artiststracks-with-metadata"
 TEST_FILEPATH = "foo.txt"
-TEST_CONTENTS = "foo\n"
+TEST_CONTENTS = "foo"
+AUTO_COMPRESSED_CONTENTS = """"shape","degrees","sides","color","date"
+"square",360,4,"blue","2024-12-17"
+"circle",360,,"red","2023-08-01"
+"triangle",180,3,"green","2022-01-05"
+"""
 
 EXPECTED_DATASET_SUBDIR = os.path.join(
     DATASETS_CACHE_SUBFOLDER, "sarahjeffreson", "featured-spotify-artiststracks-with-metadata", "versions", "2"
@@ -26,7 +30,6 @@ EXPECTED_DATASET_SUBPATH = os.path.join(
     "featured-spotify-artiststracks-with-metadata",
     "versions",
     "2",
-    TEST_FILEPATH,
 )
 
 
@@ -62,14 +65,20 @@ class TestHttpDatasetDownload(BaseTestCase):
 
     def _download_test_file_and_assert_downloaded(self, d: str, dataset_handle: str, **kwargs) -> None:  # noqa: ANN003
         dataset_path = kagglehub.dataset_download(dataset_handle, path=TEST_FILEPATH, **kwargs)
-        self.assertEqual(os.path.join(d, EXPECTED_DATASET_SUBPATH), dataset_path)
+        self.assertEqual(os.path.join(d, EXPECTED_DATASET_SUBPATH, TEST_FILEPATH), dataset_path)
+        with open(dataset_path) as dataset_file:
+            self.assertEqual(TEST_CONTENTS, dataset_file.read())
 
-        with zipfile.ZipFile(dataset_path, "r") as zip_ref:
-            if TEST_FILEPATH in zip_ref.namelist():
-                with zip_ref.open(TEST_FILEPATH) as extracted_file:
-                    contents = extracted_file.read().decode()
-
-        self.assertEqual(TEST_CONTENTS, contents)
+    def _download_test_file_and_assert_downloaded_auto_compressed(
+        self,
+        d: str,
+        dataset_handle: str,
+        **kwargs,  # noqa: ANN003
+    ) -> None:
+        dataset_path = kagglehub.dataset_download(dataset_handle, path=AUTO_COMPRESSED_FILE_NAME, **kwargs)
+        self.assertEqual(os.path.join(d, EXPECTED_DATASET_SUBPATH, AUTO_COMPRESSED_FILE_NAME), dataset_path)
+        with open(dataset_path) as dataset_file:
+            self.assertEqual(AUTO_COMPRESSED_CONTENTS, dataset_file.read())
 
     def test_unversioned_dataset_download(self) -> None:
         with create_test_cache() as d:
@@ -96,6 +105,10 @@ class TestHttpDatasetDownload(BaseTestCase):
     def test_versioned_dataset_download_with_path(self) -> None:
         with create_test_cache() as d:
             self._download_test_file_and_assert_downloaded(d, VERSIONED_DATASET_HANDLE)
+
+    def test_versioned_dataset_download_with_auto_compressed_path(self) -> None:
+        with create_test_cache() as d:
+            self._download_test_file_and_assert_downloaded_auto_compressed(d, VERSIONED_DATASET_HANDLE)
 
     def test_unversioned_dataset_download_with_force_download(self) -> None:
         with create_test_cache() as d: