+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.3.6 (December 19th, 2024)
+
+* Add `kagglehub.load_dataset(...)` ([#192](https://github.com/Kaggle/kagglehub/pull/192), [#197](https://github.com/Kaggle/kagglehub/pull/197))
+* Fix auto-compressed dataset downloads ([#194](https://github.com/Kaggle/kagglehub/pull/194), [#195](https://github.com/Kaggle/kagglehub/pull/195))
+* Update `docker-hatch` and documentation ([#193](https://github.com/Kaggle/kagglehub/pull/193))
+
+## v0.3.5 (December 10th, 2024)
+
+* Include server-side errors if present ([#183](https://github.com/Kaggle/kagglehub/pull/183))
+* Made file logging opt-out by default ([#190](https://github.com/Kaggle/kagglehub/pull/190))
+* Improved version upgrade warning ([#188](https://github.com/Kaggle/kagglehub/pull/188))
+* Add `kagglehub.notebook_output_download(...)`, only supports HTTP resolver for now. ([#184](https://github.com/Kaggle/kagglehub/pull/184))
+
 ## v0.3.4 (November 8th, 2024)
 
 * Add a naive Colab dataset cache resolver ([#177](https://github.com/Kaggle/kagglehub/pull/177))